rename package, constants moved
