--- src/main/java/org/example/AccessRoles.java
@@ -1,6 +1,0 @@
-package org.example;
-
-public enum AccessRoles {
-    ADMIN,
-    USER
-}
