--- src/main/java/org/example/App.java
@@ -1,17 +1,0 @@
-package org.example;
-
-
-import java.io.IOException;
-
-
-public class App {
-
-
-    public static void main(String[] args) throws IOException {
-        Bootstrap bs = new Bootstrap();
-        bs.init();
-    }
-
-}
-
-
