--- src/main/java/org/example/commands/TaskListCommand.java
@@ -1,33 +1,0 @@
-package org.example.commands;
-
-import org.example.entity.User;
-import org.example.service.TaskService;
-
-import java.io.BufferedReader;
-
-public class TaskListCommand extends Command {
-    private TaskService taskService;
-
-    public TaskListCommand(TaskService taskService) {
-        this.taskService = taskService;
-    }
-
-    @Override
-    public String getName() {
-        return "task-list";
-    }
-
-    @Override
-    public String getDescription() {
-        return "List of the existing tasks";
-    }
-
-    @Override
-    public void execute(BufferedReader reader, User user) {
-        if (user != null) {
-            taskService.taskList(user);
-        } else {
-            System.out.println("Log in before working");
-        }
-    }
-}
