--- src/main/java/org/example/entity/Project.java
@@ -1,18 +1,0 @@
-package org.example.entity;
-
-import lombok.Value;
-
-import java.util.Date;
-import java.util.List;
-
-@Value
-public class Project {
-    private long projectID;
-    private String projectName;
-    private String projectDescription;
-    private Date startDate;
-    private Date finishDate;
-    private List<Task> taskList;
-    private long userID;
-
-}
