--- src/main/java/org/example/entity/User.java
@@ -1,13 +1,0 @@
-package org.example.entity;
-
-import lombok.Value;
-import org.example.AccessRoles;
-
-@Value
-public class User {
-    private AccessRoles accessRoles;
-    private long userID;
-    private String userName;
-    private String md5Password;
-
-}
