+++ src/main/java/org/niks/AccessRoles.java
@@ -1,0 +1,6 @@
+package org.niks;
+
+public enum AccessRoles {
+    ADMIN,
+    USER
+}
