+++ src/main/java/org/niks/App.java
@@ -1,0 +1,17 @@
+package org.niks;
+
+
+import java.io.IOException;
+
+
+public class App {
+
+
+    public static void main(String[] args) throws IOException {
+        Bootstrap bs = new Bootstrap();
+        bs.init();
+    }
+
+}
+
+
