+++ src/main/java/org/niks/commands/Command.java
@@ -1,0 +1,17 @@
+package org.niks.commands;
+
+
+import org.niks.entity.User;
+
+import java.io.BufferedReader;
+import java.io.IOException;
+
+public abstract class Command {
+
+    public abstract String getName();
+
+    public abstract String getDescription();
+
+    public abstract void execute(BufferedReader reader, User user) throws IOException;
+}
+
