+++ src/main/java/org/niks/commands/ProjectClearCommand.java
@@ -1,0 +1,33 @@
+package org.niks.commands;
+
+import org.niks.entity.User;
+import org.niks.service.ProjectService;
+
+import java.io.BufferedReader;
+
+public class ProjectClearCommand extends Command {
+    private ProjectService projectService;
+
+    public ProjectClearCommand(ProjectService projectService) {
+        this.projectService = projectService;
+    }
+
+    @Override
+    public String getName() {
+        return "project-clear";
+    }
+
+    @Override
+    public String getDescription() {
+        return "Removes all projects";
+    }
+
+    @Override
+    public void execute(BufferedReader reader, User user) {
+        if (user != null) {
+            projectService.projectClear(user);
+        } else {
+            System.out.println("Log in before working");
+        }
+    }
+}
