+++ src/main/java/org/niks/entity/Project.java
@@ -1,0 +1,18 @@
+package org.niks.entity;
+
+import lombok.Value;
+
+import java.util.Date;
+import java.util.List;
+
+@Value
+public class Project {
+    private long projectID;
+    private String projectName;
+    private String projectDescription;
+    private Date startDate;
+    private Date finishDate;
+    private List<Task> taskList;
+    private long userID;
+
+}
