+++ src/main/java/org/niks/entity/User.java
@@ -1,0 +1,13 @@
+package org.niks.entity;
+
+import lombok.Value;
+import org.niks.AccessRoles;
+
+@Value
+public class User {
+    private AccessRoles accessRoles;
+    private long userID;
+    private String userName;
+    private String md5Password;
+
+}
