+++ src/test/java/org/niks/TestOutputs.java
@@ -1,0 +1,127 @@
+package org.niks;
+
+public class TestOutputs {
+    public static final String HELP_OUTPUT = "Welcome to the Task Manager.\nSign up, please\n" +
+            "Enter user name\n" +
+            "Enter password\n" +
+            "[User newUser created]\n" +
+            "Enter user name\n" +
+            "Enter password\n" +
+            "Welcome newUser\n" +
+            "help : Help command\n" +
+            "project-create : Creation of a project\n" +
+            "project-list : List of the existing projects\n" +
+            "project-remove : Removes a project\n" +
+            "project-clear : Removes all projects\n" +
+            "task-create : Creation of a task\n" +
+            "task-list : List of the existing tasks\n" +
+            "task-remove : Removes a task\n" +
+            "task-clear : Removes all tasks\n" +
+            "user-login : User authorization\n" +
+            "user-edit : Edit user name\n" +
+            "user-exit : Log out\n" +
+            "user-info : User info: user ID, user name\n" +
+            "password-update : Current password update\n" +
+            "user-reg : Registration of a new user\n";
+
+    public static final String PROJECT_CREATE_OUTPUT = "Welcome to the Task Manager.\nSign up, please\n" +
+            "Enter user name\n" +
+            "Enter password\n" +
+            "[User newUser created]\n" +
+            "Enter user name\n" +
+            "Enter password\n" +
+            "Welcome newUser\n" +
+            "[Enter project name]\n" +
+            "[Enter project description]\n" +
+            "[Enter starting date dd.MM.yyyy]\n" +
+            "[Enter finishing date dd.MM.yyyy]\n" +
+            "[Project newProject created]\n";
+
+    public static final String PROJECT_LIST_OUTPUT = "Welcome to the Task Manager.\nSign up, please\n" +
+            "Enter user name\n" +
+            "Enter password\n" +
+            "[User newUser created]\n" +
+            "Enter user name\n" +
+            "Enter password\n" +
+            "Welcome newUser\n" +
+            "[Enter project name]\n" +
+            "[Enter project description]\n" +
+            "[Enter starting date dd.MM.yyyy]\n" +
+            "[Enter finishing date dd.MM.yyyy]\n" +
+            "[Project newProject created]\n" +
+            "Project Name: newProject\n" +
+            "Description: Project description\n" +
+            "Start date: 30.10.2020\n" +
+            "Finish date: 31.10.2020\n" +
+            "Tasks: []\n";
+
+    public static final String TASK_CREATE_OUTPUT = "Welcome to the Task Manager.\nSign up, please\n" +
+            "Enter user name\n" +
+            "Enter password\n" +
+            "[User newUser created]\n" +
+            "Enter user name\n" +
+            "Enter password\n" +
+            "Welcome newUser\n" +
+            "[Enter project name]\n" +
+            "[Enter project description]\n" +
+            "[Enter starting date dd.MM.yyyy]\n" +
+            "[Enter finishing date dd.MM.yyyy]\n" +
+            "[Project newProject created]\n[" +
+            "Enter project to include to]\n" +
+            "[Enter task name]\n" +
+            "[Enter task description]\n" +
+            "[Enter starting date dd.MM.yyyy]\n" +
+            "[Enter finishing date dd.MM.yyyy]\n" +
+            "[Task newTask created and added to the project newProject]\n";
+
+    public static final String TASK_LIST_OUTPUT = "Welcome to the Task Manager.\nSign up, please\n" +
+            "Enter user name\n" +
+            "Enter password\n" +
+            "[User newUser created]\n" +
+            "Enter user name\n" +
+            "Enter password\n" +
+            "Welcome newUser\n" +
+            "[Enter project name]\n" +
+            "[Enter project description]\n" +
+            "[Enter starting date dd.MM.yyyy]\n" +
+            "[Enter finishing date dd.MM.yyyy]\n" +
+            "[Project newProject created]\n" +
+            "[Enter project to include to]\n" +
+            "[Enter task name]\n" +
+            "[Enter task description]\n" +
+            "[Enter starting date dd.MM.yyyy]\n" +
+            "[Enter finishing date dd.MM.yyyy]\n" +
+            "[Task newTask created and added to the project newProject]\n" +
+            "[Task newTask in the project newProject]\n" +
+            "Start date: 30.11.2020\n" +
+            "Finish date: 01.12.2020\n";
+
+    public static final String PROJECT_ISOLATION_OUTPUT = "Welcome to the Task Manager.\nSign up, please\n" +
+            "Enter user name\n" +
+            "Enter password\n" +
+            "[User newUser created]\n" +
+            "Enter user name\n" +
+            "Enter password\n" +
+            "Welcome newUser\n" +
+            "[Enter project name]\n" +
+            "[Enter project description]\n" +
+            "[Enter starting date dd.MM.yyyy]\n" +
+            "[Enter finishing date dd.MM.yyyy]\n" +
+            "[Project newProject created]\n" +
+            "Enter user name\n" +
+            "Enter password\n" +
+            "[User secondUser created]\n" +
+            "Enter user name\n" +
+            "Enter password\n" +
+            "Welcome secondUser\n" +
+            "[Enter project name]\n" +
+            "[Enter project description]\n" +
+            "[Enter starting date dd.MM.yyyy]\n" +
+            "[Enter finishing date dd.MM.yyyy]\n" +
+            "[Project secondProject created]\n" +
+            "Project Name: secondProject\n" +
+            "Description: Project description\n" +
+            "Start date: 29.12.2020\n" +
+            "Finish date: 30.12.2020\n" +
+            "Tasks: []\n";
+}
